+++ apps/web-widgets/src/app/widgets/form-widget/form-widget.component.ts
@@ -25,12 +25,15 @@ export class FormWidgetComponent
   extends ShadowRootExtendedHostComponent
   implements AfterViewInit
 {
+  /** id of the form */
   @Input() id = '626b96227ad4dd0c96f3b8a1';
   // @Input() id = '642061d1b7109549fa3035e8';
 
+  /** Reference to the right sideNav */
   @ViewChild('rightSidenav', { read: ViewContainerRef })
   rightSidenav?: ViewContainerRef;
 
+  /** boolean, whether the sidenav should be shown or not */
   public showSidenav = false;
 
   /**