+++ libs/shared/src/lib/components/widgets/summary-card/summary-card.component.ts
@@ -5,6 +5,7 @@ import {
   Input,
   OnDestroy,
   OnInit,
+  Renderer2,
   TemplateRef,
   ViewChild,
 } from '@angular/core';
@@ -36,7 +37,6 @@ export type CardT = NonNullable<SummaryCardFormT['value']['card']> &
 import { Layout } from '../../../models/layout.model';
 import { FormControl } from '@angular/forms';
 import { clone, cloneDeep, isNaN, isNil } from 'lodash';
-import { searchFilters } from '../../../utils/filter/search-filters';
 import { SnackbarService, UIPageChangeEvent } from '@oort-front/ui';
 import { Dialog } from '@angular/cdk/dialog';
 import { ResourceQueryResponse } from '../../../models/resource.model';
@@ -46,6 +46,7 @@ import { GridWidgetComponent } from '../grid/grid.component';
 import { GridService } from '../../../services/grid/grid.service';
 import { ReferenceDataService } from '../../../services/reference-data/reference-data.service';
 import { ReferenceDataQueryResponse } from '../../../models/reference-data.model';
+import searchFilters from '../../../utils/filter/search-filters';
 import filterReferenceData from '../../../utils/filter/reference-data-filter.util';
 
 /** Maximum width of the widget in column units */
@@ -122,6 +123,8 @@ export class SummaryCardComponent
   public searchControl = new FormControl('');
   /** Is scrolling */
   public scrolling = false;
+  /** Is refresh card list action */
+  private triggerRefreshCardList = false;
   /** Observer resize changes */
   private resizeObserver!: ResizeObserver;
   /** Used to reset sort options when changing display mode */
@@ -131,19 +134,24 @@ export class SummaryCardComponent
     field: string | null;
     order: string;
   } = { field: null, order: '' };
+  /** Summary card grid scroll event listener */
+  private scrollEventListener!: any;
+  /** Timeout listener for summary card scroll bind set on view switch */
+  private scrollEventBindTimeout!: NodeJS.Timeout;
 
   /** @returns Get query filter */
   get queryFilter(): CompositeFilterDescriptor {
     let filter: CompositeFilterDescriptor | undefined;
+    const skippedFields = ['id', 'incrementalId'];
     if (this.searchControl.value) {
-      const skippedFields = ['id', 'incrementalId'];
       filter = {
         logic: 'and',
         filters: [
-          { logic: 'and', filters: [this.layout?.query.filter] },
           {
             logic: 'or',
-            filters: searchFilters(
+            field: '_globalSearch',
+            operator: 'contains',
+            value: searchFilters(
               this.searchControl.value,
               this.fields,
               skippedFields
@@ -228,6 +236,7 @@ export class SummaryCardComponent
    * @param elementRef Element Ref
    * @param gridService grid service
    * @param referenceDataService Shared reference data service
+   * @param renderer Angular renderer service
    */
   constructor(
     private apollo: Apollo,
@@ -240,7 +249,8 @@ export class SummaryCardComponent
     private contextService: ContextService,
     private elementRef: ElementRef,
     private gridService: GridService,
-    private referenceDataService: ReferenceDataService
+    private referenceDataService: ReferenceDataService,
+    private renderer: Renderer2
   ) {
     super();
   }
@@ -282,12 +292,21 @@ export class SummaryCardComponent
       });
       this.resizeObserver.observe(this.elementRef.nativeElement.parentElement);
     }
+    this.bindCardsScrollListener();
+  }
+
+  /**
+   * Bind the scroll event listener to the summary cards container
+   */
+  private bindCardsScrollListener() {
     if (!this.settings.widgetDisplay?.usePagination) {
-      this.summaryCardGrid.nativeElement.addEventListener(
+      if (this.scrollEventListener) {
+        this.scrollEventListener();
+      }
+      this.scrollEventListener = this.renderer.listen(
+        this.summaryCardGrid.nativeElement,
         'scroll',
-        (event: any) => {
-          this.loadOnScroll(event);
-        }
+        (event: any) => this.loadOnScroll(event)
       );
     }
   }
@@ -295,6 +314,12 @@ export class SummaryCardComponent
   override ngOnDestroy(): void {
     super.ngOnDestroy();
     this.resizeObserver.disconnect();
+    if (this.scrollEventListener) {
+      this.scrollEventListener();
+    }
+    if (this.scrollEventBindTimeout) {
+      clearTimeout(this.scrollEventListener);
+    }
   }
 
   /**
@@ -307,6 +332,23 @@ export class SummaryCardComponent
       this.sortControl.setValue(null);
       this.onSort(null);
       this.displayMode = value;
+      if (value === 'cards') {
+        if (this.scrollEventBindTimeout) {
+          clearTimeout(this.scrollEventListener);
+        }
+        // On switching views, summary card element ref is destroyed
+        // and all events attached to it are not working as they are bind to
+        // previous element, therefor we have to set them again
+        this.scrollEventBindTimeout = setTimeout(
+          () => this.bindCardsScrollListener(),
+          0
+        );
+      } else {
+        // Clean previously attached scroll listener as the element ref is destroyed
+        if (this.scrollEventListener) {
+          this.scrollEventListener();
+        }
+      }
     }
   }
 
@@ -336,7 +378,9 @@ export class SummaryCardComponent
   private async setupDynamicCards() {
     // only one dynamic card is allowed per widget
     const card = this.settings.card;
-    if (!card) return;
+    if (!card) {
+      return;
+    }
 
     if (card.resource) {
       if (this.useAggregation) {
@@ -417,13 +461,17 @@ export class SummaryCardComponent
    * @param res Query result
    */
   private updateCards(res: any) {
-    if (!res?.data) return;
+    if (!res?.data) {
+      return;
+    }
     let newCards: any[] = [];
 
     const layoutQueryName = this.layout?.query.name;
     if (this.layout) {
       const edges = res.data?.[layoutQueryName].edges;
-      if (!edges) return;
+      if (!edges) {
+        return;
+      }
 
       newCards = edges.map((e: any) => ({
         ...this.settings.card,
@@ -442,13 +490,15 @@ export class SummaryCardComponent
       return;
     }
 
-    // scrolling enabled
-    if (!this.settings.widgetDisplay?.usePagination && this.scrolling) {
+    // update card list and scroll behavior according to the card items display
+
+    if (
+      !this.settings.widgetDisplay?.usePagination &&
+      !this.triggerRefreshCardList
+    ) {
       this.cards = [...this.cards, ...newCards];
-      this.scrolling = false;
     } else {
       this.cards = newCards;
-
       if (this.displayMode == 'cards') {
         this.summaryCardGrid.nativeElement.scroll({
           top: 0,
@@ -462,8 +512,17 @@ export class SummaryCardComponent
       'totalCount',
       0
     );
-
+    this.scrolling = false;
+    this.triggerRefreshCardList = false;
     this.loading = res.loading;
+  }
+
+  /**
+   * Refetch cards from the view.
+   */
+  public refreshCardList() {
+    this.triggerRefreshCardList = true;
+    this.dataQuery.refetch();
   }
 
   /**
@@ -720,13 +779,14 @@ export class SummaryCardComponent
    * @param e scroll event
    */
   private loadOnScroll(e: any): void {
-    if (
-      e.target.scrollHeight - (e.target.clientHeight + e.target.scrollTop) <
-      50
-    ) {
+    /** If scroll is reaching bottom of scrolling height, trigger card load */
+    const isScrollNearBottom =
+      e.target.scrollHeight - (e.target.clientHeight + e.target.scrollTop) < 50;
+    if (isScrollNearBottom) {
       if (!this.scrolling && this.pageInfo.length > this.cards.length) {
+        this.scrolling = true;
         if (this.useReferenceData) {
-          const start = this.pageInfo.pageIndex * this.pageInfo.pageSize;
+          const start = this.cards.length;
           const end = start + this.pageInfo.pageSize;
           this.cards.push(...this.sortedCachedCards.slice(start, end));
           this.scrolling = false;
@@ -738,7 +798,6 @@ export class SummaryCardComponent
               },
             })
             .then(this.updateCards.bind(this));
-          this.scrolling = true;
         }
       }
     }
@@ -855,7 +914,9 @@ export class SummaryCardComponent
       this.gridComponent.onSort(e);
     } else {
       if (this.useLayout) {
-        if (!this.dataQuery) return;
+        if (!this.dataQuery) {
+          return;
+        }
         this.dataQuery
           .refetch({
             first: this.pageInfo.pageSize,